Fix Logger singleton
