+++ TestShell/TestShell.vcxproj
@@ -139,6 +139,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ISSD.h" />
+    <ClInclude Include="Logger.h" />
     <ClInclude Include="TCResult.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
