+++ TestShell/TestShell.vcxproj.filters
@@ -50,5 +50,8 @@
     <ClInclude Include="TCResult.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
+    <ClInclude Include="Logger.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
